Changes graded status to reported
Reported makes more sense than graded. The file is being graded the
entire time. The only reason this was being used was to prevent __del__
from printing out the results a second time. Therefore it makes more
sense for the variable to be named reported since the __del__ function
is really checking if the grade has already been reported